requirements.txt
fiona==1.9.4
rasterio==1.3.8
pandas==2.0.3
openai==0.27.8
rich==13.4.2
tqdm==4.65.0
python-dotenv==1.0.0
geopandas==0.13.2
google-api-python-client==2.95.0
